requirements.txt
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
constantly==23.10.4
crispy-bootstrap5==2025.4
cryptography==45.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.1
django-crispy-forms==2.4
django-csp==4.0
django-daraja==1.3.0
django-environ==0.12.0
django-extensions==4.1
django-tailwind==4.0.1
django-widget-tweaks==1.5.0

djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-filter==24.*
drf-spectacular==0.27.*
django-cors-headers==4.*


djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-filter==24.*
drf-spectacular==0.27.*
django-cors-headers==4.*

djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
django-filter==24.*
drf-spectacular==0.27.*
django-cors-headers==4.*

geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
mysqlclient==2.2.7
paypalrestsdk==1.13.3
pillow==11.2.1

psycopg==3.2.9
psycopg-binary==3.2.9

pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.1.0
python-decouple==3.8
python-dotenv==1.1.0
redis==6.4.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==12.3.0
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1



zope.interface==7.2
psycopg[binary]==3.2.9
whitenoise==6.7.0
